add eax

+++ spin_lock_test/clean_exec_num.c
@@ -2,15 +2,15 @@
 #include <stdlib.h>
 #include <unistd.h>
 #include "shm.h"
-unsigned long *exec_proc = NULL;
+unsigned int *exec_proc = NULL;
 
-unsigned long *global_lock = NULL;
+unsigned int *global_lock = NULL;
 int main()
 {
 	int i = 0;
 
 	init_shm();
-	global_lock = (unsigned long *)get_shm();
+	global_lock = (unsigned int *)get_shm();
 	if (!global_lock) {
 		printf("get shm mem error\n");
 		return 1;
